feat: add all prayer request count by church - add GET API for this feature - Domain Adapter can adopt domain based repository

+++ src/main/java/mitl/IntoTheHeaven/adapter/in/web/dto/prayer/PrayerRequestCountByChurchResponse.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 @Getter
 @Builder
 public class PrayerRequestCountByChurchResponse {
-    private final int count;
+    private final long count;
 
-    public static PrayerRequestCountByChurchResponse from(int count) {
+    public static PrayerRequestCountByChurchResponse from(Long count) {
         return PrayerRequestCountByChurchResponse.builder().count(count).build();
     }
 }
